v1.02
fix: dropped UPX compressed binaries + recompiled with VS19 instead of VS22. Fixes most malware false positives. Fixes #4 (thanks to everyone!)

new: gamepad support (Windows, Linux)
new: added setting hints (UI)
new: ZXDB use screen$ as background during AY play
new: ZXDB include ripped AY tunes
chg: update zxdb to latest (1.0.207)
chg: centered multi-line UI notifications
chg: do not reset ZX when changing issue2/3 setting
chg: added home/end keys while zxdb browsing (thanks @korb!)
chg: get rid of large pure black areas in fullscreen mode (thanks @korb!)
chg: unzip() may accept wildcards now
fix: increases X64 compatibility. dropped AVX2 support in favor of AVX1. Fixes #4 (Windows) (thanks @korb!)
fix: UB during __builtin_clz()/__builtin_clz() ops. fixes glitches while drawing font on some old cpus (thanks @korb!)
fix: fixed osx compilation errors (thanks @imneme!)
fix: edge case where modal dialogs be hidden when rendering in fullscreen mode (Windows) (thanks @korb!)
fix: fixed tape pulse overflow in pzx loading code (thanks @imneme!)
fix: misplaced pixel in ui_notify() messages
fix: fixed compilation VS+X86 build (unsupported).

wip: partial RZX support
wip: partial SZX support

+++ src/zx_tzx.h
@@ -519,7 +519,7 @@ int pzx_load(const byte *fp, int len) {
         if( !memcmp(&tag, "PULS", 4) ) {
             for( unsigned num = bytes; num; ) {
                 uint16_t count = 1;
-                uint16_t duration = *(uint16_t*)ptr; ptr += 2; num -= 2;
+                uint32_t duration = *(uint16_t*)ptr; ptr += 2; num -= 2;
                 if( duration > 0x8000 ) {
                     count = duration & 0x7FFF;
                     duration = *(uint16_t*)ptr; ptr += 2; num -= 2;
